NFDIV-5413 Refactor to use existing object DeemedServiceJourneyOptions
