+++ src/main/java/uk/gov/hmcts/divorce/solicitor/event/page/DeemedServiceDetailsAndUploadPage.java
@@ -2,8 +2,10 @@ package uk.gov.hmcts.divorce.solicitor.event.page;
 
 import uk.gov.hmcts.divorce.common.ccd.CcdPageConfiguration;
 import uk.gov.hmcts.divorce.common.ccd.PageBuilder;
+import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
-import uk.gov.hmcts.divorce.divorcecase.model.DeemedServiceApplication;
+import uk.gov.hmcts.divorce.divorcecase.model.DeemedServiceJourneyOptions;
+import uk.gov.hmcts.divorce.divorcecase.model.InterimApplicationOptions;
 
 public class DeemedServiceDetailsAndUploadPage implements CcdPageConfiguration {
 
@@ -32,13 +34,17 @@ public class DeemedServiceDetailsAndUploadPage implements CcdPageConfiguration {
     public void addTo(PageBuilder pageBuilder) {
         pageBuilder.page("deemedServiceDetailsAndUpload")
             .pageLabel("Deemed Service App")
-            .complex(CaseData::getDeemedServiceApplication)
-            .label("deemedEvidenceLabel", DEEMED_EVIDENCE_LABEL)
-            .mandatory(DeemedServiceApplication::getDeemedEvidenceDetails)
-            .label("deemedStatementLabel", DEEMED_STATEMENT_LABEL)
-            .optional(DeemedServiceApplication::getDeemedStatement)
-            .label("deemedEvidenceDocsLabel", DEEMED_EVIDENCE_DOCS_LABEL)
-            .optional(DeemedServiceApplication::getDeemedEvidenceDocs)
+            .complex(CaseData::getApplicant1)
+                .complex(Applicant::getInterimApplicationOptions)
+                    .complex(InterimApplicationOptions::getDeemedServiceJourneyOptions)
+                        .label("deemedEvidenceLabel", DEEMED_EVIDENCE_LABEL)
+                        .mandatory(DeemedServiceJourneyOptions::getDeemedEvidenceDetails)
+                        .label("deemedStatementLabel", DEEMED_STATEMENT_LABEL)
+                        .optional(DeemedServiceJourneyOptions::getDeemedNoEvidenceStatement)
+                        .label("deemedEvidenceDocsLabel", DEEMED_EVIDENCE_DOCS_LABEL)
+                        .optional(DeemedServiceJourneyOptions::getDeemedEvidenceDocs)
+                    .done()
+                .done()
             .done();
     }
 }
